+++ src/main/java/com/Artisan/repository/ChatMessageRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.Artisan.entities.ChatMessage;
 
 
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {}
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Integer> {}
