+++ src/main/java/com/Artisan/repository/ProductRepository.java
@@ -3,7 +3,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.Artisan.entities.Product;
 
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
-	
-}
+public interface ProductRepository extends JpaRepository<Product, Integer> {}
